feat: initial lyricsfile support
