+++ Cargo.toml
@@ -29,79 +29,80 @@ clap_complete = "4.5.65"
 clap_mangen = "0.2.31"
 vergen-gitcl = { version = "9.1.0", features = ["build"] }
 
-dialoguer = { version = "0.12.0", default-features = false }
-bon = "3.8.2"
-base64 = "0.22.1"
-crossterm = { version = "0.29.0", features = ["serde", "osc52"] }
-image = "0.25.9"
-ratatui = { version = "0.30.0", features = ["all-widgets"] }
-flate2 = { version = "1.1.8" }
-ron = "0.12.0"
-derive_more = { version = "2.1.1", features = ["into_iterator", "into", "as_ref", "display", "deref", "debug"] }
-rustix = { version = "1.1.3", features = ["termios", "fs", "stdio", "process"] }
-bitflags = { version = "2.10.0", features = ["serde"] }
-flexi_logger = "0.31.8"
-chrono = { version = "0.4.43", features = ["serde"] }
-serde_with = "3.16.1"
-either = "1.15.0"
-walkdir = "2.5.0"
-which = "8.0.0"
-url = "2.5.8"
-sysinfo = "0.38.0"
-color_quant = "1.1.0"
-enum-map = "2.7.3"
-textwrap = "0.16.2"
-crossbeam = "0.8.4"
-notify-debouncer-full = "0.7.0"
-unicode-width = "=0.2.0"
-unicase = "2.9.0"
-serde_path_to_error = "0.1.20"
-serde_json = "1.0.149"
 ansi_colours = "1"
-thiserror = "2.0.18"
-parking_lot = { version = "0.12.5", features = [] }
-unicode-segmentation = "1.12.0"
-winnow = "^0.7.14"
-socket2 = { version = "0.5", features = ["all"] }
+base64 = "0.22.1"
+bitflags = { version = "2.10.0", features = ["serde"] }
+bon = "3.8.2"
+chrono = { version = "0.4.43", features = ["serde"] }
+color_quant = "1.1.0"
+console-subscriber = "0.5"
+crossbeam = "0.8.4"
+crossterm = { version = "0.29.0", features = ["serde", "osc52"] }
+ctrlc = { version = "3.5.2" }
+derive_more = { version = "2.1.1", features = ["into_iterator", "into", "as_ref", "display", "deref", "debug"] }
+dialoguer = { version = "0.12.0", default-features = false }
+either = "1.15.0"
+enum-map = "2.7.3"
+flate2 = { version = "1.1.8" }
+flexi_logger = "0.31.8"
+image = "0.25.9"
+md5 = { version = "0.8.0" }
 mlua = { version = "0.11.6", features = ["lua55", "vendored", "anyhow", "serde", "async", "send"]}
-zbus = { version = "5.14.0", features = ["tokio"] }
+notify-debouncer-full = "0.7.0"
+parking_lot = { version = "0.12.5", features = [] }
+ratatui = { version = "0.30.0", features = ["all-widgets"] }
+reqwest = { version = "0.13.2", features = ["json", "query"] }
+ron = "0.12.0"
+rustix = { version = "1.1.3", features = ["termios", "fs", "stdio", "process"] }
+serde_json = "1.0.149"
+serde_path_to_error = "0.1.20"
+serde_with = "3.16.1"
+serde_yaml_ng = "0.10.0"
+socket2 = { version = "0.5", features = ["all"] }
+sysinfo = "0.38.0"
+textwrap = "0.16.2"
+thiserror = "2.0.18"
 tokio = { version = "1.49.0", features = ["full", "tracing"] }
 tokio-util = "0.7.18"
 tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-console-subscriber = "0.5"
-reqwest = { version = "0.13.2", features = ["json", "query"] }
-md5 = { version = "0.8.0" }
-ctrlc = { version = "3.5.2" }
 tracing-appender = "0.2"
 tracing-log = "0.2"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+unicase = "2.9.0"
+unicode-segmentation = "1.12.0"
+unicode-width = "=0.2.0"
+url = "2.5.8"
+walkdir = "2.5.0"
+which = "8.0.0"
+winnow = "^0.7.14"
+zbus = { version = "5.14.0", features = ["tokio"] }
 
 [workspace.lints.clippy]
 pedantic = { level = "deny", priority = -1 }
-single_match = "allow"
-type_complexity = "allow"
-module_name_repetitions = "allow"
-unused_self = "allow"
-unnested_or_patterns = "allow"
-match_same_arms = "allow"
-manual_let_else = "allow"
-needless_return = "allow"
-zero_sized_map_values = "allow"
-too_many_lines = "allow"
-match_single_binding = "allow"
-struct_field_names = "allow"
-redundant_closure_for_method_calls = "allow"
-unwrap_used = "deny"
+cast_lossless = "allow"
 cast_possible_truncation = "allow"
 cast_precision_loss = "allow"
-cast_lossless = "allow"
 cast_sign_loss = "allow"
-must_use_candidate = "allow"
-missing_errors_doc = "allow"
-return_self_not_must_use = "allow"
 duration_suboptimal_units = "allow"
+manual_let_else = "allow"
+match_same_arms = "allow"
+match_single_binding = "allow"
+missing_errors_doc = "allow"
+module_name_repetitions = "allow"
+must_use_candidate = "allow"
+needless_return = "allow"
+redundant_closure_for_method_calls = "allow"
+return_self_not_must_use = "allow"
+single_match = "allow"
+struct_field_names = "allow"
+too_many_lines = "allow"
+type_complexity = "allow"
+unnested_or_patterns = "allow"
+unused_self = "allow"
+unwrap_used = "deny"
+zero_sized_map_values = "allow"
 
 [workspace.lints.rust]
-unused_macros = "allow"
 deprecated = "allow"
 unused_crate_dependencies = "deny"
+unused_macros = "allow"
