+++ rmpc/Cargo.toml
@@ -15,44 +15,45 @@ workspace = true
 [dependencies]
 rmpc-mpd = { path = "../rmpc-mpd" }
 rmpc-shared = { path = "../rmpc-shared" }
-clap = { workspace = true }
-dialoguer = { workspace = true }
+ansi_colours = { workspace = true }
 anyhow = { workspace = true }
-bon = { workspace = true }
 base64 = { workspace = true }
-crossterm = { workspace = true }
-image = { workspace = true }
-ratatui = { workspace = true }
-strum = { workspace = true }
-flate2 = { workspace = true }
-itertools = { workspace = true }
-ron = { workspace = true }
-derive_more = { workspace = true }
-rustix = { workspace = true }
 bitflags = { workspace = true }
-log = { workspace = true }
-flexi_logger = { workspace = true }
+bon = { workspace = true }
 chrono = { workspace = true }
-serde = { workspace = true }
-serde_with = { workspace = true }
+clap = { workspace = true }
+color_quant = { workspace = true }
+crossbeam = { workspace = true }
+crossterm = { workspace = true }
+derive_more = { workspace = true }
+dialoguer = { workspace = true }
 either = { workspace = true }
+enum-map = { workspace = true }
+flate2 = { workspace = true }
+flexi_logger = { workspace = true }
+image = { workspace = true }
+itertools = { workspace = true }
+log = { workspace = true }
+notify-debouncer-full = { workspace = true }
+parking_lot = { workspace = true }
+ratatui = { workspace = true }
+ron = { workspace = true }
+rustix = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+serde_path_to_error = { workspace = true }
+serde_with = { workspace = true }
+serde_yaml_ng = { workspace = true }
+strum = { workspace = true }
+sysinfo = { workspace = true }
+textwrap = { workspace = true }
+thiserror = { workspace = true }
+unicase = { workspace = true }
+unicode-segmentation = { workspace = true }
+unicode-width = { workspace = true }
+url = { workspace = true }
 walkdir = { workspace = true }
 which = { workspace = true }
-url = { workspace = true }
-sysinfo = { workspace = true }
-color_quant = { workspace = true }
-enum-map = { workspace = true }
-textwrap = { workspace = true }
-crossbeam = { workspace = true }
-notify-debouncer-full = { workspace = true }
-unicode-width = { workspace = true }
-unicase = { workspace = true }
-serde_path_to_error = { workspace = true }
-serde_json = { workspace = true }
-ansi_colours = { workspace = true }
-thiserror = { workspace = true }
-parking_lot = { workspace = true }
-unicode-segmentation = { workspace = true }
 winnow = { workspace = true }
 
 [dev-dependencies]
