+++ rmpc/src/ctx.rs
@@ -31,7 +31,7 @@ use crate::{
     shared::{
         events::ClientRequest,
         keys::KeyResolver,
-        lrc::{Lrc, LrcIndex},
+        lrc::{Lrc, LrcIndex, lyricsfile_parser},
         macros::{status_error, status_warn},
         mpd_client_ext::{MpdClientExt, PartitionedOutput},
         mpd_query::MpdQuerySync,
@@ -273,29 +273,48 @@ impl Ctx {
         self.status.song
     }
 
-    pub(crate) fn find_current_lyrics_path(&self) -> Option<PathBuf> {
+    pub(crate) fn find_current_lyrics_path(&self) -> Option<LyricsResult> {
         let song = self.current_song()?;
         let lyrics_dir = self.config.lyrics_dir.as_ref()?;
-        let path = crate::shared::lrc::get_lrc_path(lyrics_dir, &song.file)
+
+        // Priority: lyricsfile.yaml > .lrc > lrc_index
+        let lrc_path = crate::shared::lrc::get_lyricsfile_path(lyrics_dir, &song.file)
             .ok()
             .filter(|p| p.is_file())
-            .or_else(|| self.lrc_index.find_entry(song).map(|(path, _)| path.to_path_buf()));
+            .map(LyricsResult::Lyricsfile)
+            .or_else(|| {
+                crate::shared::lrc::get_lrc_path(lyrics_dir, &song.file)
+                    .ok()
+                    .filter(|p| p.is_file())
+                    .map(LyricsResult::Lrc)
+            })
+            .or_else(|| {
+                self.lrc_index
+                    .find_entry(song)
+                    .map(|(path, _)| path.to_path_buf())
+                    .map(LyricsResult::Index)
+            });
 
         let artist = song.metadata.get("artist").map(|v| v.last())?;
         let title = song.metadata.get("title").map(|v| v.last())?;
         let album = song.metadata.get("album").map(|v| v.last());
-        match &path {
+        match &lrc_path {
             Some(path) => log::debug!(artist, title, album; "Lyrics found at {}", path.display()),
             None => log::debug!(artist, title, album; "No lyrics found"),
         }
 
-        path
+        lrc_path
     }
 
-    pub(crate) fn find_lrc(&self) -> Result<Option<(PathBuf, Lrc)>> {
-        let Some(path) = self.find_current_lyrics_path() else { return Ok(None) };
-        let lrc = std::fs::read_to_string(&path)?.parse()?;
-        Ok(Some((path, lrc)))
+    pub(crate) fn find_lrc(&self) -> Result<Option<(LyricsResult, Lrc)>> {
+        let Some(result) = self.find_current_lyrics_path() else { return Ok(None) };
+        let lrc = match &result {
+            LyricsResult::Lrc(path) | LyricsResult::Index(path) => {
+                std::fs::read_to_string(path)?.parse()?
+            }
+            LyricsResult::Lyricsfile(path) => lyricsfile_parser::parse(path)?,
+        };
+        Ok(Some((result, lrc)))
     }
 
     pub(crate) fn song_stickers(&self, uri: &str) -> Option<&HashMap<String, String>> {
@@ -341,6 +360,22 @@ impl Ctx {
 
     pub(crate) fn stickers(&self) -> &HashMap<String, HashMap<String, String>> {
         &self.stickers
+    }
+}
+
+pub enum LyricsResult {
+    Lrc(PathBuf),
+    Lyricsfile(PathBuf),
+    Index(PathBuf),
+}
+
+impl LyricsResult {
+    pub fn display(&self) -> std::path::Display<'_> {
+        match self {
+            LyricsResult::Lrc(path) => path.display(),
+            LyricsResult::Lyricsfile(path) => path.display(),
+            LyricsResult::Index(path) => path.display(),
+        }
     }
 }
